PCA Implemented
MIrar data_cleaning file IMPORTANTE
