+++ Case1/data_cleaning.py
@@ -3,6 +3,7 @@ import numpy as np
 from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import LabelEncoder
 from sklearn.pipeline import Pipeline
+from sklearn.decomposition import PCA
 
 def rem_cat(df):
     df['C_ 1'].replace(['I','H','K','G','J'],[0,1,2,3,4],inplace=True)
@@ -12,52 +13,18 @@ def rem_cat(df):
     df['C_ 5'].replace(['I','H','K','G','J'],[0,1,2,3,4],inplace=True)
 
 
-class MultiColumnLabelEncoder:
-    def __init__(self,columns = None):
-        self.columns = columns # array of column names to encode
-
-    def fit(self, X,y=None):
-        return self # not relevant here
-
-    def transform(self,X):
-        '''
-        Transforms columns of X specified in self.columns using
-        LabelEncoder(). If no columns specified, transforms all
-        columns in X.
-        '''
-        output = X.copy()
-        if self.columns is not None:
-            for col in self.columns:
-                output[col] = LabelEncoder().fit_transform(output[col])
-        else:
-            for colname,col in output.iteritems():
-                output[colname] = LabelEncoder().fit_transform(col)
-        return output
-
-    def fit_transform(self,X,y=None):
-        return self.fit(X, y).transform(X)
-
 def load_data():
     X = pd.read_csv(r"case1Data.txt", sep=', ', engine='python')
     X_Xn = pd.read_csv(r"case1Data_Xnew.txt",  sep=', ', engine='python')
     return X, X_Xn
 
-def standarize(df,method):
-    if method == 'standard':
-        df = (df - df.mean()) / df.std()
-    elif method == 'minmax':
-        df = (df - df.min()) / (df.max() - df.min())
-    elif method == 'maxabs':
-        df = df / df.abs().max()
-    elif method == 'robust':
-        df = (df - df.median()) / (df.quantile(0.75) - df.quantile(0.25))
-    else:
-        raise ValueError('Invalid standardization method')
-    return df
-
 def transform_data(df_train, df_test, fill_method, std_method):
     # Encode categorical variables
-    imputer_numeric = SimpleImputer(missing_values=np.nan, strategy=fill_method)
+    if type(std_method) is int:
+    	pp = "mean"
+    else:
+    	pp = fill_method
+    imputer_numeric = SimpleImputer(missing_values=np.nan, strategy=pp)
     imputer_categorical = SimpleImputer(missing_values=np.nan, strategy='most_frequent')
     
     df_num = df_train._get_numeric_data()
@@ -84,6 +51,18 @@ def transform_data(df_train, df_test, fill_method, std_method):
         c = df_num.quantile(0.25)
         df_num_std = (df_num - a) / (b - c)
         df_tt_num_std = (df_tt_num - a) / (b-c)
+    elif (type(std_method) is int) and (std_method > 0):
+        a = df_num.mean()
+        b = df_num.std()
+        df_num_std = (df_num - a) / b
+        df_tt_num_std = (df_tt_num - a) / b
+        
+        aux1 = imputer_numeric.fit_transform(df_num_std)
+        aux1_tt = imputer_numeric.transform(df_tt_num_std)
+        
+        pca = PCA(n_components = std_method)
+        df_num_std = pca.fit_transform(aux1)
+        df_tt_num_std = pca.transform(aux1_tt)    	
     else:
         raise ValueError('Invalid standardization method')
     
@@ -95,25 +74,14 @@ def transform_data(df_train, df_test, fill_method, std_method):
     df_tt_cat = df_test.drop(df_tt_num, axis =1)
     rem_cat(df_cat)
     rem_cat(df_tt_cat)
-    
-    """
-    df_cat_encoded = MultiColumnLabelEncoder(columns=df_cat.columns).fit_transform(df_cat)
-    df_tt_cat_encoded = MultiColumnLabelEncoder(columns=df_tt_cat.columns).fit_transform(df_cat)
-    
-    mask = df_cat.isnull()
-    mask_tt = df_tt_cat.isnull()
-    df_cat_temp = df_cat_encoded.where(~mask)
-    df_tt_cat_temp = df_tt_cat_encoded.where(~mask_tt)
-	"""
+
     df_cat = imputer_categorical.fit_transform(df_cat)
     df_tt_cat = imputer_categorical.transform(df_tt_cat)
 
-    df_trans = pd.DataFrame(np.concatenate([df_num_final, df_cat], axis=1), columns=df_train.columns)
-    df_tt_trans = pd.DataFrame(np.concatenate([df_tt_num_final, df_tt_cat], axis=1), columns=df_test.columns)
+    df_trans = pd.DataFrame(np.concatenate([df_num_final, df_cat], axis=1))#, columns=df_train.columns)
+    df_tt_trans = pd.DataFrame(np.concatenate([df_tt_num_final, df_tt_cat], axis=1))#, columns=df_test.columns)
 
     return df_trans.values,df_tt_trans.values
-
-
